% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cup_mfg_overall.R
\docType{data}
\name{cup_mfg_overall}
\alias{cup_mfg_overall}
\title{NASCAR Cup Series Manufacturer Overall Statistics}
\format{
A data frame with 361 rows and 13 variables:
\describe{
\item{manufacturer}{chr: Manufacturer name}
\item{mfg_overall_races}{num: Total number of races for the manufacturer}
\item{mfg_overall_wins}{num: Total number of wins for the manufacturer}
\item{mfg_overall_win_pct}{num: Win percentage for the manufacturer}
\item{mfg_overall_top_5}{num: Total number of top 5 finishes for the manufacturer}
\item{mfg_overall_top_10}{num: Total number of top 10 finishes for the manufacturer}
\item{mfg_overall_top_20}{num: Total number of top 20 finishes for the manufacturer}
\item{mfg_overall_cars_raced}{num: Total number of cars raced by the manufacturer}
\item{mfg_overall_car_win_pct}{num: Win percentage at the car level for the manufacturer}
\item{mfg_overall_avg_start}{num: Average start position for the manufacturer}
\item{mfg_overall_avg_finish}{num: Average finish position for the manufacturer}
\item{mfg_overall_laps_led}{num: Total number of laps led by the manufacturer}
\item{mfg_overall_avg_laps_led}{num: Average number of laps led per race for the manufacturer}
}
}
\source{
\url{https://www.nascar.com/news/nascar-cup-series/}

\url{https://www.driveraverages.com/nascar/nascar-stats.php}

\url{https://www.racing-reference.info/nascar-cup-series-stats/}

\url{https://en.wikipedia.org/wiki/NASCAR_Cup_Series}
}
\usage{
cup_mfg_overall
}
\description{
A collection of overall statistics for manufacturers in the NASCAR Cup Series. The data includes overall totals and averages for various performance metrics.
}
\examples{
data(cup_mfg_overall)
}
\keyword{datasets}
