% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_driver_data}
\alias{filter_driver_data}
\title{Filter race data based on selected driver}
\usage{
filter_driver_data(race_data, the_driver)
}
\arguments{
\item{race_data}{The data frame returned from \code{selected_series_data()}.}

\item{the_driver}{The name of the driver returned from \code{find_driver()}.}
}
\value{
A filtered data frame of \code{class()} "tbl_df" containing race results
of the selected driver. A win column is added for summary statistics.
}
\description{
Internal helper function to filter race data based on the selected driver.
}
\examples{
\dontrun{
filter_driver_data(cup_series, 'Jimmie Johnson')
}
}
\keyword{internal}
