% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfinity_owner_career.R
\docType{data}
\name{xfinity_owner_career}
\alias{xfinity_owner_career}
\title{NASCAR Xfinity Series Owner Career Statistics}
\format{
A data frame with 1241 rows and 13 variables:
\describe{
\item{owner}{chr: Owner name}
\item{owner_overall_races}{num: Total number of races for the owner in their career}
\item{owner_overall_wins}{num: Total number of wins for the owner in their career}
\item{owner_overall_win_pct}{num: Win percentage for the owner in their career}
\item{owner_overall_top_5}{num: Total number of top 5 finishes for the owner in their career}
\item{owner_overall_top_10}{num: Total number of top 10 finishes for the owner in their career}
\item{owner_overall_top_20}{num: Total number of top 20 finishes for the owner in their career}
\item{owner_overall_cars_raced}{num: Total number of cars raced by the owner in their career}
\item{owner_overall_car_win_pct}{num: Win percentage at the car level for the owner in their career}
\item{owner_overall_avg_start}{num: Average start position for the owner in their career}
\item{owner_overall_avg_finish}{num: Average finish position for the owner in their career}
\item{owner_overall_laps_led}{num: Total number of laps led by the owner in their career}
\item{owner_overall_avg_laps_led}{num: Average number of laps led per race for the owner in their career}
}
}
\source{
\url{https://www.nascar.com/news/nascar-xfinity-series/}

\url{https://www.driveraverages.com/nascar_xfinityseries/}

\url{https://www.racing-reference.info/nascar-xfinity-series-stats/}

\url{https://en.wikipedia.org/wiki/NASCAR_Xfinity_Series}
}
\usage{
xfinity_owner_career
}
\description{
A collection of career statistics for owners in the NASCAR Xfinity Series. The data includes career totals and averages for various performance metrics.
}
\examples{
data(xfinity_owner_career)
}
\keyword{datasets}
