% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_driver}
\alias{find_driver}
\title{Find Driver Matches}
\usage{
find_driver(search_term, data = NULL, max_results = 5, interactive = TRUE)
}
\arguments{
\item{search_term}{Character string to search for}

\item{data}{Tibble containing NASCAR race data}

\item{max_results}{Maximum number of matches to return}
}
\value{
Character vector of matching driver names
}
\description{
Find Driver Matches
}
