% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{nascaR.data-package}
\alias{nascaR.data}
\alias{nascaR.data-package}
\title{nascaR.data: NASCAR Race Data}
\description{
A collection of NASCAR race, driver, owner and manufacturer data across the three major NASCAR divisions: NASCAR Cup Series, NASCAR Xfinity Series, and NASCAR Craftsman Truck Series. The curated data begins with the 1949 season and extends through the end of the 2024 season. Explore race, season, or career performance for drivers, teams, and manufacturers throughout NASCAR's history. Data was sourced with permission from DriverAverages.com.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://azimuth-project.tech/nascaR.data/}
}

}
\author{
\strong{Maintainer}: Kyle Grealis \email{kylegrealis@icloud.com} (\href{https://orcid.org/0000-0002-9223-8854}{ORCID})

Other contributors:
\itemize{
  \item Nick Triplett \email{nicholas.triplett@siu.edu} [contributor]
  \item Gabriel Odom \email{gabriel.odom@fiu.edu} (\href{https://orcid.org/0000-0003-1341-4555}{ORCID}) [contributor]
}

}
\keyword{internal}
