% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selected_series_data}
\alias{selected_series_data}
\title{Filter race data by series}
\usage{
selected_series_data(series)
}
\arguments{
\item{series}{A string specifying the race series. Must be one of
'cup', 'xfinity', 'truck', or 'all'.}
}
\value{
A tibble containing race results for the specified series.
}
\description{
Filter race data by series
}
\keyword{internal}
